ecuhojaruta/requirements.txt
Django==2.2.4
django-crispy-forms==1.7.2
django-discover-runner==1.0
django-role-permissions==2.2.1
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
